finish up the application
here we tried the best practice when meeting IO and CPU bound operations. such that we should always put async method close to the closest event handler method. we also should only back to the main thread when have to.  We have achieved using UWP application to show and interact with methods that uses IO and CPU bound operations. where we should run tasks in other threads from thread pool (to reduce thread creating overhead). we also should let them run async to not block the main UI thread while they are running. we implemented UWP API such that they do not contains async methods which makes it possible to consumers to decide how to run the code to follow the best practice